Change Shader to use floats in shader because texture coordinates are between 0 and 1 for spritesheets

+++ core/include/Sprite.h
@@ -27,13 +27,13 @@ namespace Core
 		std::string getName();
 		std::string getFileName();
 
-		int* getTextureCoordinates();
+		float* getTextureCoordinates();
 
 	protected:
 		std::string mSpriteName;
 		std::string mFileName;
 
-		int mTextureCoordinates[8];
+		float mTextureCoordinates[8];
 
 		glm::vec2 mImageDimensions;
